Switch to U8g2 display and secure web API
Migrate OLED code from Adafruit_SSD1306 to U8g2 (display.cpp/.h, wlms_v2.0.ino), add U8g2 include to VSCode config. Refactor display APIs and drawing logic (better fonts, tank fill, icons) and adjust volume type. Harden web dashboard: remove ArduinoJson dependency, implement manual JSON parsing, HMAC-SHA256 using mbedtls, constant-time compare, nonce/timestamp replay protection, IP restriction, rate limiting and a new handleSystemCtrl handler for /api/7f3a9c_motor_ctrl (web_dash.cpp). Remove embedded web credentials from config.h. Add fsm_controller prototypes, integrate WiFi event handling and LoRa support (wlms_master.ino) including LoRa init and packet receive logic, plus various small cleanups and formatting tweaks.

+++ WLMS_MASTER/wlms_master.ino/display.h
@@ -4,4 +4,4 @@
 
 void initDisplay();
 void welcomeScreen();
-void updateDisplay(uint8_t level, uint16_t voltage, uint8_t volume, const char* dateTime, bool isMotorOn = false, bool isWifiConnected = false, bool isMainsCut = false);
+void updateDisplay(uint8_t level, uint16_t voltage, uint16_t volume, const char* dateTime, bool isMotorOn = false, bool isWifiConnected = false, bool isMainsCut = false);
